Make database a little less magic
This also helps to support passing in and potentially parsing any type of query,
rather than only model specific ones. This is all speculation on what I might
want to query for, but _hopefully_ these patterns will turn out to be useful in
the future when I need to query for subsets of data rather than full things.
(though I'm not really sure how I'll model that, may need to build in some
default parsers based on types, dunno)