Omg you can actually share inventory now
This is all a little gross, but it WORKS! And I think it's secure with all the
rules set up and what now. *wipes sweat*.

Wow this was a significant amount of work for this seemingly simple feature,
but it is really helping me to better understand how to structure firebase data
and query for it. (Though some of these queries are pretty gnarly, RxSwift is
the savior here for making it managable)