Upgrade to Swift 3
Wow. This was a lot of work. Xcode couldn't even handle compiling everything
with errors so it would freeze up. The only way I was able to get it to start
working was to go file by file and let it partially compile it and fix.

A few interesting things:

- `with` seems to no longer work without completx type annotations which kind of
  defeats it's purpose.
- Switched all deps to Cocoapods since I was having odd issues when using
  Carthage. Have to specify all the versions explicitly which is kind of gross but
  not many have an official Swift 3 release yet so *shrugs*.
- Keychain access group enabled since Firebase Auth needs it to work in the sim
- Remove Firebase Auth UI since they don't support Swift 3 yet :(, for now
  I switched to using an alert prompt for sign in, which means people can't sign
  up, so this will need to be repaired for releasing to testers.