Inventory sharing working with the security rules!
This wasn't working due to the security rules specifying that you can't write
to public user unless you have already set that you're sharing your inventory
with them. These still aren't really strict enough, we should probably have
something that says you can only add or remove yourself, but hopefully we can
do that later. This is a step in that direction though.

It included adding an `observeSave` method since the ordering of these writes
matters. I chose to prefix with `observe` since also calling it `save` would
conflict with the existing methods and cause you to have to include type
annotations all over to use it.

I also added the `select` method on `Dictionary` to pull out specific keys with
the ability of a "partial" match, this part was added for pulling out values
when we have a fully qualified path, like "PublicUsers/12/name" and being able
to say we want "name". It's pretty naive but does the trick. Perhaps a better
future one would only match on `endsWith`? Dunno.