Fix removing a shared inventory partner
Because of the structure of the seurity rules, removing wasn't working due to
the user no longer marked as sharing their inventory, so it prevented a write to
the other user.

This reordering make sure that the authed user's `inventorySharedWith` is added
first when adding a sharer, and removed last when removing one.