Fix sign out
Sign out wasn't working because we were using the wrong form of `drive`. This
broke in the Swift 3 update.

Also add individual actions to update the in memory state of the FIRuser and
User. Without this I think we probably were keeping things in memory that we
shouldn't when a user signs out. It also makes it easier to understand.
Basically the issue was if you passed in `.none`, it wouldn't reset the value,
and we want it to.