Denormalize line items and remove firebase sync!
Denormalize line items by adding product info/price into the nested object. This
allows you to still see information about the product even if your inventory
access is removed later. This also make sure that history is maintained at the
time it was purchased, if a price changes in the future, we don't want the price
in history to be updated. It should be a snapshot of the information at the time.

Doing this allowed us to remove the firebase sync since we don't need quick
reference to products anymore! WOOO!