Fix products not loading when no shared inventory
The `.next` event from `productIDs` was never being fired when an empty array of
IDs was being passed in. Because of this, it being used in `combineLatest` was
never firing it's next event, thus never finishing loading the products.

Since the list of IDs is static, we can simply return a `.just([])` so that it
fires.