Don't overwrite existing values on users when sharing list
The users who had access to the list would get their "Lists" overridden with a
only the new list because we were updating it at the root.

We add a `deepFlatten` method to dictionaries and use that to flatten it into
a flat keypath like `["Users/1234/Lists/542": true]` to correct this.

Use `take(1)` instead of single on create since `.single()` prints an error if
we get more than one, and in this case we expect it to change, but we only want
one, so `take` is the correct choice.

Also handle case where products or users haven't been set yet.