Delete a list
Allow a user to delete a list by swiping it and tapping delete.

This works, but comes with an issue mentioned in a TODO.

With the way I plan on setting up security rules, this would inadvertently
remove a user's write access to a product if they happen to also be on another
list that requires access to the product.