Spike setting up realm
Using realm's new backend was a potential alternative to firebase that supported
relationships and would make things a bit easier, but it doesn't support any
sort of security other than a user owns what they create, so the sharing
wouldn't work.

I considered setting it up with all public data to start but since I have no
ETA or idea of what type of security they may offer in the future, that puts me
in a vulnerable spot for the future so I want to avoid it.

For now I'll probably continue using Firebase and deal with the difficulties it
creates because of no relationship support and complex security rules. It might
also be good for me to write up the data model I'd like, and reach out to
someone if they know how I could set that up. In particular, the list sharing
seems problematic and I haven't been able to determine a way to set up the
security rules correctly.